split widgets (rotate + crop)
