+++ CMakeLists.txt
@@ -67,7 +67,8 @@ set(XPANO_SOURCES
   "xpano/gui/panels/warning_pane.cc"
   "xpano/gui/pano_gui.cc"
   "xpano/gui/shortcut.cc"
-  "xpano/gui/widgets/widgets.cc"
+  "xpano/gui/widgets/drag.cc"
+  "xpano/gui/widgets/rotate.cc"
   "xpano/pipeline/options.cc"
   "xpano/pipeline/stitcher_pipeline.cc"
   "xpano/utils/config.cc"
