+++ xpano/gui/panels/preview_pane.cc
@@ -117,82 +117,6 @@ void Overlay(const widgets::RotationWidget& widget, const utils::RectPVf& image,
   }
 }
 
-template <typename... TEdge>
-constexpr int Select(TEdge... edges) {
-  return (static_cast<int>(edges) + ...);
-}
-
-void SelectMouseCursor(const widgets::RotationWidget& widget) {
-  const int mouse_cursor_selector = std::accumulate(
-      widget.rotation.edges.begin(), widget.rotation.edges.end(), 0,
-      [](int sum, const widgets::Edge& edge) {
-        return sum + (edge.mouse_close || edge.dragging
-                          ? static_cast<int>(edge.type)
-                          : 0);
-      });
-
-  switch (mouse_cursor_selector) {
-    case Select(EdgeType::kRoll, EdgeType::kHorizontal):
-      [[fallthough]];
-    case Select(EdgeType::kHorizontal):
-      ImGui::SetMouseCursor(ImGuiMouseCursor_ResizeNS);
-      break;
-    case Select(EdgeType::kRoll, EdgeType::kVertical):
-      [[fallthough]];
-    case Select(EdgeType::kVertical):
-      ImGui::SetMouseCursor(ImGuiMouseCursor_ResizeEW);
-      break;
-    case Select(EdgeType::kRoll, EdgeType::kHorizontal, EdgeType::kVertical):
-      [[fallthough]];
-    case Select(EdgeType::kHorizontal, EdgeType::kVertical):
-      ImGui::SetMouseCursor(ImGuiMouseCursor_ResizeAll);
-      break;
-    case Select(EdgeType::kRoll):
-      ImGui::SetMouseCursor(ImGuiMouseCursor_Hand);
-      break;
-    default:
-      ImGui::SetMouseCursor(ImGuiMouseCursor_Arrow);
-      break;
-  }
-}
-
-void SelectMouseCursor(const widgets::DraggableWidget& crop) {
-  const int mouse_cursor_selector = std::accumulate(
-      crop.edges.begin(), crop.edges.end(), 0, [](int sum, const widgets::Edge& edge) {
-        return sum + (edge.mouse_close || edge.dragging
-                          ? static_cast<int>(edge.type)
-                          : 0);
-      });
-
-  switch (mouse_cursor_selector) {
-    case Select(EdgeType::kTop):
-      [[fallthrough]];
-    case Select(EdgeType::kBottom):
-      ImGui::SetMouseCursor(ImGuiMouseCursor_ResizeNS);
-      break;
-    case Select(EdgeType::kLeft):
-      [[fallthrough]];
-    case Select(EdgeType::kRight):
-      ImGui::SetMouseCursor(ImGuiMouseCursor_ResizeEW);
-      break;
-    case Select(EdgeType::kBottom, EdgeType::kRight):
-      [[fallthrough]];
-    case Select(EdgeType::kTop, EdgeType::kLeft): {
-      ImGui::SetMouseCursor(ImGuiMouseCursor_ResizeNWSE);
-      break;
-    }
-    case Select(EdgeType::kBottom, EdgeType::kLeft):
-      [[fallthrough]];
-    case Select(EdgeType::kTop, EdgeType::kRight): {
-      ImGui::SetMouseCursor(ImGuiMouseCursor_ResizeNESW);
-      break;
-    }
-    default:
-      ImGui::SetMouseCursor(ImGuiMouseCursor_Arrow);
-      break;
-  }
-}
-
 }  // namespace
 
 PreviewPane::PreviewPane(backends::Base* backend) : backend_(backend) {
